Conflicts:
	Living Life.v11.suo
	Living Life/bin/Debug/Living Life.exe
	Living Life/bin/Debug/Living Life.pdb
	Living Life/obj/Debug/Living Life.csproj.GenerateResource.Cache
	Living Life/obj/Debug/Living Life.exe
	Living Life/obj/Debug/Living Life.pdb

+++ Living Life/Player.cs
@@ -11,8 +11,8 @@ namespace Living_Life
         public int age;
         public int wife;
         public int children;
-        public float income;
-        public float savings;
+        public int income;
+        public int savings;
         public int educationLevel;
         public int schoolMonths;
         public Property car;
